Refactor apps section to reduce component complexity.
