+++ ui/src/components/app-card/app-card.css
