+++ ui/src/utils/jupyterhub.ts
@@ -1,5 +1,6 @@
 import { JhApp, JhService, JhServiceFull } from '@src/types/jupyterhub';
 import { JhData } from '@src/types/jupyterhub.ts';
+import { UserState } from '@src/types/user';
 import { DEFAULT_APP_THUMBNAIL, DEFAULT_PINNED_SERVICES } from './constants';
 
 export const getJhData = (): JhData => {
@@ -106,4 +107,47 @@ export const getAppStatus = (app: JhApp): string => {
   } else {
     return 'Unknown';
   }
+};
+
+export const filterAndSortApps = (
+  data: UserState,
+  currentUser: UserState,
+  searchValue: string,
+  ownershipValue: string,
+  frameworkValues: string[],
+  sortByValue: string,
+) => {
+  const searchToLower = searchValue.toLowerCase();
+  const ownershipType =
+    ownershipValue === 'Owned by me'
+      ? 'mine'
+      : ownershipValue === 'Shared with me'
+        ? 'shared'
+        : 'all';
+
+  // Get Apps based on ownership type and search value
+  const apps = getApps(data, ownershipType, currentUser?.name ?? '')
+    .filter(
+      (app) =>
+        app.name.toLowerCase().includes(searchToLower) ||
+        app.description?.toLowerCase().includes(searchToLower) ||
+        app.framework?.toLowerCase().includes(searchToLower),
+    )
+    .filter((app) => {
+      if (frameworkValues.length > 0) {
+        return frameworkValues.includes(app.framework);
+      }
+      return true;
+    });
+
+  // Sort Apps based on sort value
+  apps.sort((a, b) => {
+    if (sortByValue === 'Recently modified') {
+      return a.last_activity > b.last_activity ? -1 : 1;
+    } else if (sortByValue === 'Name: A-Z') {
+      return a.name > b.name ? 1 : -1;
+    }
+    return a.name > b.name ? -1 : 1;
+  });
+  return apps;
 };
